requirements.txt
blinker==1.6.2
boto3==1.34.13
botocore==1.34.13
cffi==1.16.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.5
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
Pillow==10.1.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
Werkzeug==2.3.6
zipp==3.16.1
